Added the create product feature.
Got the values of variables in request body to create a product constant. Specified the resource path. CreateRequest object is initialized and sent it to createProduct function as a parameter. Specified the path of the product and the product set. Created addRequest with these paths and sent it to addProductToProductSet function as a parameter.